Config object format is wrong
For me, what was in the readme didn't work. I was getting errors that svgo was trying to loop over an obejct. From reading the source, it looks like you expect the plugins passed in, not a config object containing a `plugins` property. I don't know how the tests are passing but it works for me now when I pass the plugins array in as the param.